+++ History.md
@@ -4,6 +4,7 @@ Unpublished
   * export Gpio constructor as exports.Gpio
   * high precision notifications with Notifier
   * hardware pwm support
+  * renamed analogWrite to pwmWrite
 
 0.0.8 / Nov 08 2015
 ===================